Edit end move State pattern, main example
