+++ patterns/java/src/refactoring_guru/patterns/state/example/Demo.java
@@ -1,7 +1,7 @@
-package refactoring_guru.patterns.state;
+package refactoring_guru.patterns.state.example;
 
-import refactoring_guru.patterns.state.ui.Player;
-import refactoring_guru.patterns.state.ui.UI;
+import refactoring_guru.patterns.state.example.ui.Player;
+import refactoring_guru.patterns.state.example.ui.UI;
 
 public class Demo {
     public static void main(String[] args) {
