+++ patterns/java/src/refactoring_guru/patterns/state/example/ui/UI.java
@@ -1,4 +1,4 @@
-package refactoring_guru.patterns.state.ui;
+package refactoring_guru.patterns.state.example.ui;
 
 import javax.swing.*;
 import java.awt.*;
